split various.test.js in auth.test.js and cookie.test.js
